refactor: rename file to lowercase
