+++ packages/app/src/systems/Core/components/PermissionCard/mocks/index.ts
@@ -1,1 +1,1 @@
-export * from './PERMISSIONS';
+export * from './permissions';
